refactor(monorepo): complete migration to pnpm monorepo with package restructuring
- Finalize migration steps by resolving all import path and compilation issues
- Update README and project requirements to reflect monorepo structure
- Adjust package.json dependencies and devDependencies for shared, indesign-server, and illustrator-server
- Refactor imports in all packages to use workspace shared package references
- Add zod and @types/node dependencies where needed
- Fix telemetry and experimental imports and exports
- Prepare servers for independent build and start commands
- Ensure successful build and run of both InDesign and Illustrator servers

This commit completes the monorepo refactoring, enabling better code organization, independent package management, and improved build and run workflows.
