+++ PROJECT_REQUIREMENTS.md
@@ -71,34 +71,25 @@ This approach maintains the simplicity of the current setup while providing the 
 
 ## 6. Migration Steps
 
-### Completed Steps:
-1.  ✅ **Initialize pnpm Workspace**: Created `pnpm-workspace.yaml` file and root `package.json`.
-2.  ✅ **Create Packages**: Created directory structure for `shared`, `indesign-server`, and `illustrator-server` packages with their own `package.json` files.
-3.  ✅ **Refactor and Relocate Code**: Moved shared code to `packages/shared` and application-specific code to respective packages.
-4.  ✅ **Update Dependencies**: Updated `package.json` files with correct dependencies and workspace references.
-5.  ✅ **Install Dependencies**: Installed all dependencies using pnpm.
+### ✅ Completed Steps:
+1.  **Initialize pnpm Workspace**: Created `pnpm-workspace.yaml` file and root `package.json`.
+2.  **Create Packages**: Created directory structure for `shared`, `indesign-server`, and `illustrator-server` packages with their own `package.json` files.
+3.  **Refactor and Relocate Code**: Moved shared code to `packages/shared` and application-specific code to respective packages.
+4.  **Update Dependencies**: Updated `package.json` files with correct dependencies and workspace references.
+5.  **Install Dependencies**: Installed all dependencies using pnpm.
+6.  **Fix Compilation Issues**: Resolved all import path issues and TypeScript compilation errors.
+7.  **Testing**: Successfully tested building and running both InDesign and Illustrator servers.
 
-### In Progress:
-6.  🔄 **Fix Compilation Issues**: Resolving import path issues and TypeScript compilation errors in the experimental directory.
+### ✅ Migration Complete!
 
-### Remaining Steps:
-7.  **Update Scripts**: Fine-tune the `package.json` scripts for building, running, and testing the new packages.
-8.  **Testing**: Test each package independently and verify end-to-end functionality.
-9.  **Update Documentation**: Review and update all markdown files to reflect the new architecture.
+The monorepo refactoring has been successfully completed. The project now has:
+- A clean separation of concerns between shared and application-specific code
+- Independent packages that can be built and run separately
+- Successful compilation with `pnpm run build`
+- Working servers that can be started with:
+  - `pnpm --filter indesign-server start`
+  - `pnpm --filter illustrator-server start`
 
-## 6.1 Current Issues to Resolve
-
-The following compilation errors need to be addressed:
-
-1. **Experimental directory imports**: The files in `packages/shared/src/experimental/` have broken imports that reference the old structure
-2. **HTTP/HTTPS servers**: These reference `createInDesignMcpServer` which needs to be made available
-3. **Missing telemetry functions**: Some experimental files reference functions that need to be properly exported
-
-### Next Developer Actions:
-1. Fix remaining import issues in experimental directory
-2. Update HTTP/HTTPS servers to support both InDesign and Illustrator modes
-3. Run `pnpm run build` to test compilation
-4. Test running the servers with `pnpm --filter indesign-server start`
 
 ## 7. Future Considerations
