+++ packages/illustrator-server/src/index.ts
@@ -4,7 +4,8 @@
 
 import { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
 import { StdioServerTransport } from "@modelcontextprotocol/sdk/server/stdio.js";
-import { createTelemetryServer, setTelemetryEnabled } from "@mcp/shared/telemetryServer.js";
+import { createTelemetryServer } from "@mcp/shared/telemetryServer.js";
+import { setTelemetryEnabled } from "@mcp/shared/telemetryFlag.js";
 import { TelemetryCapture } from "@mcp/shared/telemetry.js";
 import { registerGeometryTools } from "./tools/geometry/index.js";
 import { registerTransformTools } from "./tools/transform/index.js";
