+++ packages/shared/src/experimental/evolutionary/runner.ts
@@ -9,8 +9,8 @@ import { spawn } from 'child_process';
 import * as path from 'path';
 import * as fs from 'fs/promises';
 import * as os from 'os';
-import { TelemetryCapture, TelemetrySession } from "../../telemetry.js.js';
-import { TelemetryPersistence } from "../../telemetry.jsPersistence.js';
+import { TelemetryCapture, TelemetrySession } from "../../telemetry.js";
+import { TelemetryPersistence } from "../../telemetryPersistence.js";
 import { LayoutMetrics, ComparisonResult } from '../../types.js';
 import { SubAgentExecutor, createSubAgentExecutor } from './subAgentExecutor.js';
 import { getMcpBridge, McpBridge } from './mcpBridge.js';
